Renamed SendGridTestEmail to TestEmail
There's no need to specify what the provider is
