--- api/ExpressedRealms.Email/SendGridTestEmail/ISendGridEmail.cs
@@ -1,6 +1,0 @@
-namespace ExpressedRealms.Email.SendGridTestEmail;
-
-public interface ISendGridEmail
-{
-    Task SendTestEmail();
-}
