+++ api/ExpressedRealms.Email/TestEmail/ITestEmail.cs
@@ -1,0 +1,6 @@
+namespace ExpressedRealms.Email.TestEmail;
+
+public interface ITestEmail
+{
+    Task SendTestEmail();
+}
